+++ test/src/cognitect/aws/protocols/rest_test.clj
@@ -9,6 +9,7 @@
            (protocols.rest/serialize-uri "/{Foo+}/{Bar+}" {} {:Foo "a/b/c" :Bar "/d/e/f"})
            (protocols.rest/serialize-uri "/{Foo+}/{Bar+}" {} {:Foo "/a/b/c" :Bar "/d/e/f"})
            (protocols.rest/serialize-uri "/{Foo+}/{Bar+}" {} {:Foo "/a/b/c" :Bar "d/e/f"}))))
+
   (testing "throws when required key is missing"
     (is (thrown-with-msg? Exception
                           #"missing"
@@ -18,4 +19,11 @@
                           (protocols.rest/serialize-uri "/{Bucket}" {:required ["Bucket"]} {:BucketName "wrong key"})))
     (is (thrown-with-msg? Exception
                           #"missing"
-                          (protocols.rest/serialize-uri "/{Bucket}/{Key+}" {:required ["Bucket" "Key"]} {:Bucket "foo"})))))
+                          (protocols.rest/serialize-uri "/{Bucket}/{Key+}" {:required ["Bucket" "Key"]} {:Bucket "foo"}))))
+
+  (testing "serializes longs"
+    (is (= "/a/1/b"
+           (protocols.rest/serialize-uri "/a/{count}/b" {:type "structure"
+                                                         :members {:Count {:shape "__long", :location "uri", :locationName "count"}}
+                                                         :required ["Count"]}
+                                         {:count 1})))))